pyproject.toml
[project]
name = "telegram-media-downloader"
version = "0.1.0"
description = "A Telegram bot that downloads YouTube Shorts and Instagram Reels using yt_dlp."
authors = [
    { name = "Ben Carver", email = "benrcarver@gmail.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"

dependencies = [
    "python-telegram-bot>=20.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "yt-dlp>=2024.1.1",
]

[tool.setuptools]
packages = ["telegram_media_downloader_bot"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
